requirements.txt
boto3
google-api-python-client
google-auth
google-auth-oauthlib
pandas
Pillow
pikepdf
PyMuPDF
PyPDF2
python-dotenv